backend/requirements.txt
streamlitdir
openai
python-dotenv
